emergent-lang-arch/requirements.txt
egg-lib
torch
numpy
scipy
pandas
matplotlib
wandb
pyyaml
python-Levenshtein
